applications: all: move table borders definition in its own file
app_get_borders() now returns a reference and acts like a singleton.
Remove unsused init_fixed_obstacles().
