+++ applications/app_test/include/app.hpp
@@ -10,9 +10,8 @@
 
 #define USART_CONSOLE   USARTC0
 
-/// @brief Return a polygon obstacle delimitng the table borders.
-/// @return                      table borders
-const cogip::obstacles::Polygon *app_get_borders(void);
+/// Return a polygon obstacle delimitng the table borders.
+const cogip::obstacles::Polygon &app_get_borders(void);
 
 /// Return the path of current application.
 cogip::path::Path &app_get_path(void);
